+++ otlet/cli/options.py
@@ -2,11 +2,14 @@ import sys
 from argparse import ArgumentParser
 from .arguments import *
 
-if "releases" not in sys.argv:
+if "releases" in sys.argv:
+    RELEASES_ARGUMENT_LIST["package"] = PACKAGE_ARGUMENT
+elif "download" in sys.argv:
+    DOWNLOAD_ARGUMENTS_LIST["package"] = PACKAGE_ARGUMENT
+    DOWNLOAD_ARGUMENTS_LIST["package_version"] = PACKAGE_VERSION_ARGUMENT
+else:
     ARGUMENT_LIST["package"] = PACKAGE_ARGUMENT
     ARGUMENT_LIST["package_version"] = PACKAGE_VERSION_ARGUMENT
-else:
-    RELEASES_ARGUMENT_LIST["package"] = PACKAGE_ARGUMENT
 
 class OtletArgumentParser(ArgumentParser):
     def __init__(self):
@@ -17,10 +20,11 @@ class OtletArgumentParser(ArgumentParser):
         )
         self.arguments = ARGUMENT_LIST
         self.releases_arguments = RELEASES_ARGUMENT_LIST
+        self.download_arguments = DOWNLOAD_ARGUMENTS_LIST
         for key, arg in self.arguments.items():
             self.add_argument(*arg["opts"], **self.without_keys(arg, ["opts"]), dest=key)
-        pro = [_ for _ in ["releases", "--help", "-h"] if _ in sys.argv] # i did this because i can
-        if pro:
+        scinit = [_ for _ in ["releases", "download", "--help", "-h"] if _ in sys.argv] # i did this because i can
+        if scinit:
             self.subparsers = self.add_subparsers(parser_class=ArgumentParser, metavar="[ sub-commands ]")
             self.releases_subparser = self.subparsers.add_parser(
                 "releases",
@@ -28,8 +32,16 @@ class OtletArgumentParser(ArgumentParser):
                 help="List releases for a specified package",
                 epilog="(c) 2022-present Noah Tanner, released under the terms of the MIT License",
             )
+            self.download_subparser = self.subparsers.add_parser(
+                "download",
+                description="dl",
+                help="dl",
+                epilog="(c) 2022-present Noah Tanner, released under the terms of the MIT License",
+            )
             for key,arg in self.releases_arguments.items():
                 self.releases_subparser.add_argument(*arg["opts"], **self.without_keys(arg, ["opts"]), dest=key)
+            for key,arg in self.download_arguments.items():
+                self.download_subparser.add_argument(*arg["opts"], **self.without_keys(arg, ["opts"]), dest=key)
 
     def without_keys(self, d, keys):
         return {x: d[x] for x in d if x not in keys}